requirements-dev.txt
# styling
black

# distribution
build
twine

# testing
PyYAML
